add file upload utility to handle media uploads to Supabase and update FileUploadTest component
