+++ package.json
@@ -16,7 +16,8 @@
     "react-dom": "^19.0.0",
     "react-hot-toast": "^2.5.2",
     "react-icons": "^5.5.0",
-    "react-router-dom": "^7.2.0"
+    "react-router-dom": "^7.2.0",
+    "timestamp": "^0.0.1"
   },
   "devDependencies": {
     "@eslint/js": "^9.19.0",
